backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
httpx==0.25.2